adjusted the enumerables
