--- list.rb
@@ -1,14 +1,0 @@
-require_relative './enumerable'
-
-class MyList
-  include Enumerable
-
-  def initialize(*list)
-    @list = list
-  end
-
-  def each
-    (0...@list.length).each do |item|
-      yield @list[item]
-    end
-  end
